Add basic functionality
Add basic functionality so that the revenues can be added with a date and an amount of the revenue. The revenues will be shown in a table with a sum over the table for the current month. There is also a progress bar so that everyone can the how much percent of the amount goal is reached for this month. Added revenues can be removed with the trashcan button.